Remove unnecesary print statements
